Factor out firebase.tsx.
