+++ lib/firebase.tsx
@@ -9,6 +9,7 @@ import {
   User,
 } from "firebase/auth";
 import React, { useCallback, useContext, useEffect, useState } from "react";
+import { AuthContext } from "./auth-context";
 
 const DEFAULT_APP_CONFIG: FirebaseOptions = {
   apiKey: "AIzaSyAV1kkVvSKEicEa8rLke9o_BxYBu1rb8kw",
@@ -19,36 +20,6 @@ const DEFAULT_APP_CONFIG: FirebaseOptions = {
   appId: "1:26787182745:web:e4fbd9439b9279fe966008",
   measurementId: "G-JHKRSK1PR6",
 };
-
-/**
- * Generic interface for authentication.
- */
-export interface AuthContext {
-  /**
-   * The currently logged-in user. This will be
-   * null if the user isn't logged in, otherwise it will
-   * be their name.
-   */
-  loggedInUser: string | null;
-
-  /**
-   * The name of the authentication provider, e.g. "GitHub",
-   * or null if auth is disabled.
-   */
-  providerName: string | null;
-
-  /**
-   * If authentication failed for some reason, this will
-   * be a string describing the error.
-   */
-  error?: string;
-
-  /** Begin the login UI flow. */
-  login(): void;
-
-  /** Log out the user. */
-  logout(): void;
-}
 
 type FirebaseGithubAuthState = {
   auth: Auth;
@@ -116,10 +87,3 @@ export const FirebaseGithubAuthProvider: React.FC<{}> = ({ children }) => {
 
   return <AuthContext.Provider value={context} children={children} />;
 };
-
-export const AuthContext = React.createContext<AuthContext>({
-  loggedInUser: null,
-  providerName: null,
-  login() {},
-  logout() {},
-});
